Build dummy journalbeat on non-windows (#8810)
Journald and headers are only available on linux. With this change we
only build a stub-journalbeat and allow unit tests which do not require
linux dependencies (similar to winlogbeat)

(cherry picked from commit 967d417a28019b3e9555500c3a77b3129767e4d0)
