+++ journalbeat/cmd/instance/metrics.go
@@ -15,6 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+//+build linux,cgo
+
 package instance
 
 import (
