+++ journalbeat/reader/journal.go
@@ -15,6 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+//+build linux,cgo
+
 package reader
 
 import (
@@ -34,27 +36,6 @@ import (
 	"github.com/elastic/beats/libbeat/common"
 	"github.com/elastic/beats/libbeat/logp"
 )
-
-const (
-	// LocalSystemJournalID is the ID of the local system journal.
-	LocalSystemJournalID = "LOCAL_SYSTEM_JOURNAL"
-)
-
-// Config stores the options of a reder.
-type Config struct {
-	// Path is the path to the journal file.
-	Path string
-	// Seek specifies the seeking stategy.
-	// Possible values: head, tail, cursor.
-	Seek string
-	// MaxBackoff is the limit of the backoff time.
-	MaxBackoff time.Duration
-	// Backoff is the current interval to wait before
-	// attemting to read again from the journal.
-	Backoff time.Duration
-	// Matches store the key value pairs to match entries.
-	Matches []string
-}
 
 // Reader reads entries from journal(s).
 type Reader struct {
